bug fix: add breadcrumb url to cache context
It turns out the current page in the preprocess doesn't flush when
moving to other pages. So, breadcrumbs work when you go to one page,
but once you start clicking the active page is stuck on previous pages.
This should fix it by basically clearing the cache when preprocessing.

See DS-338