fix: switch to keyup listeners rather than keydown where needed
One of the techniques for satisfying the newer pointer cancellation requirement is to
use keyup rather than keydown for key-specific functions (on Esc press, on arrow press, etc.).
Some code was still using keydown, so this adjusts to use keyup.

The exception is the lightbox focus trap, where keydown is needed to simulate tabbing to the close button/CTA link either way.