refactor: remove extraneous <div> markup from statistics
The statistics on the sbdc dev site were sounding choppy with VoiceOver--tried to use the technique I did with the quick search to make the statistic sound like one paragraph, but didn't have luck (maybe it only works with position: absolute elements...). Still thought it couldn't hurt to make the stats more semantic with <p> tags, so this adjusts the templates that way. Still works in the Drupal block and PL for me.

See WTR-275