DS-391 fix: remove breadcrumb block from revision nodes
The revision feature of Drupal nodes don't agree with the way we need to customize the breadcrumb items so the current page title and URL show up. Since revisions are more of an admin thing, we can just remove the block for now so we don't have to keep setting up exclusions. It'd be even better to find a way to make the preprocess work with revision versions of pages, but that seems a lot harder so that could be a long term feature.

Resolves DS-391